# app/static/swagger.yaml
swagger: "2.0"

info:
  title: "Community Fridge"
  version: 0.1.0

paths:
  /fridges:
    get:
      summary: "List all fridges, optionally sort by distance to a location"
      parameters:
        - name: latitude
          in: query
          type: number
        - name: longitude
          in: query
          type: number
      responses:
        200:
          schema:
            type: object
            properties:
              fridge_id:
                type: integer
  /fridges/{fridge_id}/items:
    get:
      summary: "List all items in a fridge"
      description: ""
      parameters:
        - name: fridge_id
          in: path
          required: true
          type: integer
      responses:
        200:
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
    post:
      summary: "Put a new item to the fridge"
      description: ""
      parameters:
        - name: fridge_id
          in: path
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              user:
                $ref: "#/definitions/User"
              item_name:
                type: string
      responses:
        200:
          schema:
            $ref: "#/definitions/ActionId"
  /items/{item_id}:
    delete:
      summary: "Get an item from the fridge"
      description: ""
      parameters:
        - name: item_id
          in: path
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              user:
                $ref: "#/definitions/User"
      responses:
        200:
          schema:
            $ref: "#/definitions/ActionId"
  /actions/{action_id}/picture:
    put:
      summary: "Upload picture for an action"
      description: ""
      consumes:
        - image/jpeg
      parameters:
        - name: action_id
          in: path
          required: true
          type: integer
        - name: body
          in: body
          required: true
          type: string
          description: "The picture in JPEG format encoded in Base64"
      responses:
        204:
          description: No Content

definitions:
  User:
    type: object
    properties:
      name:
        type: string
      password:
        type: string
  ActionId:
    type: object
    properties:
      action_id:
        type: integer